Cleaner, mostly-functional version of dynamic panes!
Using views where I had been using crazy helpers to no effect.  Got rid
of the weird array of strings and using real Ember templates.  Most
importantly, updates to the data are actually updating the view.  The
missing piece is it won't render panes that haven't had anything
assigned to them, nor is there anything in the markup to show which slot
in the pane that the templates are being rendered inside of.